+++ mitra_models/src/groups/queries.rs
@@ -186,9 +186,13 @@ mod tests {
         accounts::test_utils::create_test_user,
         activitypub::constants::AP_PUBLIC,
         database::test_utils::create_test_database,
+        groups::{
+            helpers::join_private_group,
+            test_utils::create_test_remote_group,
+        },
         posts::{
             queries::create_post,
-            types::{PostCreateData, PostContext},
+            types::{PostCreateData, PostContext, Visibility},
         },
         profiles::{
             queries::{create_profile, delete_profile},
@@ -270,16 +274,13 @@ mod tests {
     async fn test_get_group_timeline() {
         let db_client = &mut create_test_database().await;
         let account = create_test_user(db_client, "user").await;
-        let group_data = {
-            let mut group_data = ProfileCreateData::remote_for_test(
-                "group",
-                "groups.example",
-                "https://groups.example/123",
-            );
-            group_data.actor_type = ActorType::Group;
-            group_data
-        };
-        let group = create_profile(db_client, group_data).await.unwrap();
+        let group = create_test_remote_group(
+            db_client,
+            "group",
+            "groups.example",
+            "https://groups.example/123",
+            false, // public group
+        ).await;
         let author = create_test_remote_profile(
             db_client,
             "author",
@@ -306,6 +307,64 @@ mod tests {
         ).await.unwrap();
         assert_eq!(posts.len(), 1);
         assert_eq!(posts[0].id, post.id);
+    }
+
+    #[tokio::test]
+    #[serial]
+    async fn test_get_group_timeline_private_group() {
+        let db_client = &mut create_test_database().await;
+        let group = create_test_remote_group(
+            db_client,
+            "group",
+            "groups.example",
+            "https://groups.example/123",
+            true,
+        ).await;
+        assert!(group.is_private_group());
+        let author = create_test_remote_profile(
+            db_client,
+            "author",
+            "social.example",
+            "https://social.example/users/1",
+        ).await;
+        join_private_group(db_client, author.id, group.id).await.unwrap();
+        let post_data = PostCreateData {
+            context: PostContext::Top {
+                group_id: Some(group.id),
+                object_id: Some("https://social.example/contexts/123".to_owned()),
+                audience: Some("https://groups.example/123/followers".to_owned()),
+            },
+            visibility: Visibility::Group,
+            object_id: Some("https://social.example/posts/123".to_owned()),
+            ..PostCreateData::for_test()
+        };
+        let post = create_post(
+            db_client,
+            author.id,
+            post_data,
+        ).await.unwrap();
+
+        let account = create_test_user(db_client, "user").await;
+        join_private_group(db_client, account.id, group.id).await.unwrap();
+        let posts = get_group_timeline(
+            db_client,
+            group.id,
+            account.id,
+            None,
+            20,
+        ).await.unwrap();
+        assert_eq!(posts.len(), 1);
+        assert_eq!(posts[0].id, post.id);
+
+        let non_member_account = create_test_user(db_client, "other").await;
+        let posts = get_group_timeline(
+            db_client,
+            group.id,
+            non_member_account.id,
+            None,
+            20,
+        ).await.unwrap();
+        assert_eq!(posts.len(), 0);
     }
 
     #[tokio::test]