+++ mitra_models/src/groups/test_utils.rs
@@ -3,6 +3,18 @@ use apx_core::crypto::{
     rsa::generate_weak_rsa_key,
 };
 
+use crate::{
+    database::DatabaseClient,
+    profiles::{
+        queries::create_profile,
+        types::{
+            ActorType,
+            DbActorProfile,
+            ProfileCreateData,
+        },
+    },
+};
+
 use super::types::GroupCreateData;
 
 impl GroupCreateData {
@@ -16,4 +28,23 @@ impl GroupCreateData {
             ed25519_secret_key: generate_weak_ed25519_key(),
         }
     }
+}
+
+pub async fn create_test_remote_group(
+    db_client: &mut impl DatabaseClient,
+    username: &str,
+    hostname: &str,
+    actor_id: &str,
+    is_private: bool,
+) -> DbActorProfile {
+    let mut group_data = ProfileCreateData::remote_for_test(
+        username,
+        hostname,
+        actor_id,
+    );
+    group_data.actor_type = ActorType::Group;
+    group_data.manually_approves_followers = is_private;
+    let group = create_profile(db_client, group_data).await.unwrap();
+    group.check_consistency().unwrap();
+    group
 }